requirements.txt
streamlit
gspread
google-auth
pandas
numpy
yfinance
plotly